# src/main/resources/swagger.yaml
swagger: "2.0"
info:
  description: "API para gerenciamento de clientes"
  version: "1.0.0"
  title: "Swagger Clientes"
host: "localhost:8080"
basePath: "/v1"
tags:
- name: "clientes"
  description: "Operações relacionadas a cliente"
schemes:
- "http"
paths:
  /clientes:
    post:
      tags:
      - "clientes"
      summary: "Salva cliente"
      description: "Salva cliente body"
      operationId: "salvaCliente"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Salva cliente com os dados recebidos"
        required: true
        schema:
          $ref: "#/definitions/ClienteDto"
      responses:
        "201":
          description: "Criado com sucesso"
          schema:
            $ref: "#/definitions/ClienteDto"
        "400":
          description: "Bad request"
        "500":
          description: "Erro interno servidor"
    get:
      tags:
      - "clientes"
      summary: "Lista clientes por filtro"
      operationId: "buscaClientes"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "nome"
        type: "string"
      - in: "query"
        name: "idade"
        type: "integer"
        format: "int32"
      - in: "query"
        name: "sexo"
        type: "string"
      - in: "query"
        name: "pagina"
        type: "integer"
        format: "int32"
        required: true
      responses:
        "200":
          description: "Sucesso"
          schema:
            $ref: "#/definitions/ClienteResponseDto"
        "400":
          description: "Bad request"
        "500":
          description: "Erro interno servidor"
  /clientes/{id}:
    get:
      tags:
      - "clientes"
      summary: "Retorna cliente por id"
      description: "Retorna um unico cliente"
      operationId: "getCliente"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Sucesso"
          schema:
            $ref: "#/definitions/ClienteDto"
        "404":
          description: "Cliente nao encontrado"
        "500":
          description: "Erro interno servidor"
    put:
      tags:
      - "clientes"
      summary: "Atualiza cliente"
      description: "Atualiza cliente e retorna um unico cliente"
      operationId: "atualizaCliente"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Atualiza cliente com os dados recebidos"
        required: true
        schema:
          $ref: "#/definitions/ClienteDto"
      responses:
        "200":
          description: "Alterado com sucesso"
          schema:
            $ref: "#/definitions/ClienteDto"
        "404":
          description: "Cliente nao encontrado"
        "500":
          description: "Erro interno servidor"
    delete:
      tags:
      - "clientes"
      summary: "Deleta cliente por id"
      operationId: "deletaCliente"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "204":
          description: "Deletado com sucesso"
        "404":
          description: "Cliente nao encontrado"
        "500":
          description: "Erro interno servidor"
definitions:
  ClienteDto:
    type: "object"
    required : [ "nome", "idade", "sexo" ]
    properties:
      id:
        type: "integer"
        format: "int64"
      nome:
        type: "string"
      idade:
        type: "integer"
        format: "int32"
      sexo:
        $ref: '#/definitions/SexoEnum'
  SexoEnum:
      type: "string"
      description: "Estados sexo enum"
      enum:
        - "MASCULINO"
        - "FEMININO"
  ClienteResponseDto:
    type: "object"
    properties:
      totalRegistros:
        type: "integer"
        format: "int64"
      clientes:
        type: "array"
        items:
          $ref: '#/definitions/ClienteDto'